+++ CHANGELOG.md
@@ -4,3 +4,4 @@
   * Support for aliases
   * Annotation fields with description
   * Filter by @Target
+* Annotation selection engine